perf(colab): bf16 on all CUDA, bigger gen batch, 1.7B teacher toggle
- bf16 instead of fp16/fp32 for teacher gen (gen_sft) and cache (distill):
  fp32 exponent range = sampling-safe, half the memory, fits bigger
  teachers/batches on T4.
- notebook: GEN_BATCH 16->32 (uses more GPU), add SmolLM2-1.7B-Instruct
  toggle (same 49k vocab -> student size unchanged).

+++ src/train/distill.py
@@ -76,7 +76,7 @@ def cache_teacher_logits(args):
     out_dir = Path(args.out)
     out_dir.mkdir(parents=True, exist_ok=True)
     device = "cuda" if torch.cuda.is_available() else "cpu"
-    dtype = torch.float16 if device == "cuda" else torch.float32
+    dtype = torch.bfloat16 if device == "cuda" else torch.float32
     print(f"teacher: {args.teacher}  device={device}  dtype={dtype}")
 
     tok = AutoTokenizer.from_pretrained(args.teacher, trust_remote_code=True)
